Add files via upload
Scrabble game on a 15x15 table unsing C. The program had to count the points of the two players as they were making moves and decide the winner. The table had some distinct squares with the property to double/triple the letter's value or the word's value. The input letters of each player were given as a set of a token with a value and color of the square to be put on, a row and a column position. The word forming on the table was possible by elongation and intersection.